+++ includes/camera.h
@@ -5,12 +5,12 @@ typedef struct 	s_cam
 {
 	t_vec	*pos;
 	t_vec	*direction;
-	float 	fov;
+	double 	fov;
 	struct s_cam *next;
 }				t_cam;
 
-t_cam 	*cam_default(t_vec *pos, t_vec *direction, float fov);
-void 	cam_list(t_cam **cam_list, t_vec *pos, t_vec *direction, float fov);
+t_cam 	*cam_default(t_vec *pos, t_vec *direction, double fov);
+void 	cam_list(t_cam **cam_list, t_vec *pos, t_vec *direction, double fov);
 t_vec	*cam_dir(t_cam *cam);
 t_vec	*cam_up(t_vec *right, t_vec *dir);
 t_vec	*cam_right(t_vec *dir);
