+++ includes/plane.h
@@ -6,12 +6,12 @@ typedef struct	s_pl
 	t_vec		*crd;
 	t_vec		*nrmd;
 	int			color;
-	float		distance;
+	double		distance;
 	struct s_pl	*next;
 }				t_pl;
 
 t_pl	*pl_default(t_vec *crd, t_vec *nrmd, int color);
 void	pl_list(t_pl **pl_list, t_vec *crd, t_vec *nrmd, int color);
-float	pl_intersection(t_vec *pos, t_vec *ray, t_pl *pl);
+double	pl_intersection(t_vec *pos, t_vec *ray, t_pl *pl);
 
 #endif
