+++ includes/square.h
@@ -5,14 +5,14 @@ typedef struct	s_sq
 {
 	t_vec		*center;
 	t_vec		*nrmd;
-	float		side;
+	double		side;
 	int			color;
-	float		distance;
+	double		distance;
 	struct s_sq	*next;
 }				t_sq;
 
-t_sq	*sq_default(t_vec *center, t_vec *nrmd, float side, int color);
-void	sq_list(t_sq **sq_list, t_vec *center, t_vec *nrmd, float side, int
+t_sq	*sq_default(t_vec *center, t_vec *nrmd, double side, int color);
+void	sq_list(t_sq **sq_list, t_vec *center, t_vec *nrmd, double side, int
 color);
 
 #endif
